Unify e2e tests and pass all with Shared Kubernetes cluster
