+++ .github/workflows/kubernetes.yml
@@ -51,23 +51,10 @@ jobs:
         with:
           # Port forwarding sometimes dies, which makes all requests timeout
           # Which is why we need retries
-          max_attempts: 5
+          max_attempts: 10
           timeout_minutes: 45
-          command: kind get clusters | xargs -I {} kind delete cluster --name {} && cargo test --locked -p linera-service --features scylladb,kubernetes --test wasm_end_to_end_tests -- kubernetes
-      - name: Destroy the kind clusters from Wasm e2e
-        if: always()
-        shell: bash
-        run: >-
-          kind get clusters | xargs -I {} kind delete cluster --name {}
-      - name: Run e2e test
-        uses: nick-fields/retry@v2
-        with:
-          # Port forwarding sometimes dies, which makes all requests timeout
-          # Which is why we need retries
-          max_attempts: 5
-          timeout_minutes: 45
-          command: kind get clusters | xargs -I {} kind delete cluster --name {} && cargo test --locked -p linera-service --features scylladb,kubernetes --test end_to_end_tests -- test_end_to_end_multiple_wallets::kubernetes
-      - name: Destroy the kind clusters from e2e
+          command: kind get clusters | xargs -I {} kind delete cluster --name {} && cargo test --locked -p linera-service --features scylladb,kubernetes --test end_to_end_tests -- kubernetes
+      - name: Destroy the kind clusters
         if: always()
         shell: bash
         run: >-
